+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Improved [Relationship Queries](https://developer.salesforce.com/docs/atlas.en-us.soql_sosl.meta/soql_sosl/sforce_api_calls_soql_relationships.htm) support for results returned by `dataApi.query("...")`
+  so related relationship queries on a `QueriedRecord` can be accessed ([#444](https://github.com/forcedotcom/sf-fx-runtime-nodejs/pull/444))
 
 ## [0.13.0] - 2022-11-08
 - Moved SDK types from [`sf-fx-sdk-nodejs`](https://github.com/forcedotcom/sf-fx-sdk-nodejs) into runtime ([#435](https://github.com/forcedotcom/sf-fx-runtime-nodejs/pull/435))